+++ spec/grape/middleware/formatter_spec.rb
@@ -282,4 +282,14 @@ describe Grape::Middleware::Formatter do
       end
     end
   end
+
+  context 'send file' do
+    let(:app) { ->(_env) { [200, {}, @body] } }
+
+    it 'returns Grape::Uril::SendFileReponse' do
+      @body = Grape::Util::FileResponse.new('file')
+      env = { 'PATH_INFO' => '/somewhere', 'HTTP_ACCEPT' => 'application/json' }
+      expect(subject.call(env)).to be_a(Grape::Util::SendfileResponse)
+    end
+  end
 end